Fixing some formatting with admin creds and the function declaration
