+++ dabblePlusPuzzle/dabblePlusIndex.php
@@ -40,33 +40,33 @@
 
     include("../includes/innerNav.php");
 
-    if(!isset($_SESSION['loggedin']) || $_SESSION['loggedin'] == false) {
-      ?>
+    if(!is_logged_in()) {
+        ?>
+        
+        
+          <div class="right-content">
+            <div class="container">
       
+            <h3 style = "color: red;">Please log in to view this page</h3>
       
-        <div class="right-content">
-          <div class="container">
-    
-          <h3 style = "color: red;">Please log in to view this page</h3>
-    
+            </div>
           </div>
-        </div>
+        
+        <?php
+      } else if (!is_admin()) {
+        ?>
+        
+        
+          <div class="right-content">
+            <div class="container">
       
-      <?php
-    } else if (!isset($_SESSION['role']) || $_SESSION['role'] != 'admin') {
-      ?>
+            <h3 style = "color: red;">Admin privileges are required to view this page</h3>
       
-      
-        <div class="right-content">
-          <div class="container">
-    
-          <h3 style = "color: red;">Admin privileges are required to view this page</h3>
-    
+            </div>
           </div>
-        </div>
-      
-      <?php
-    } else {
+        
+        <?php
+      } else {
 ?>
 <!DOCTYPE html PUBLIC '-//W3C//DTD XHTML 1.0 Transitional//EN''http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd'>
 <html xmlns='http://www.w3.org/1999/xhtml' xml:lang='en' lang='en'>
