+++ otherPages/admin.php
@@ -5,6 +5,34 @@
 
   include("../includes/innerNav.php");
   
+  if(!is_logged_in()) {
+		?>
+		
+		
+		  <div class="right-content">
+			<div class="container">
+	  
+			<h3 style = "color: red;">Please log in to view this page</h3>
+	  
+			</div>
+		  </div>
+		
+		<?php
+	  } else if (!is_admin()) {
+		?>
+		
+		
+		  <div class="right-content">
+			<div class="container">
+	  
+			<h3 style = "color: red;">Admin privileges are required to view this page</h3>
+	  
+			</div>
+		  </div>
+		
+		<?php
+	  } else {
+
  ?>
 
 <div class="right-content">
@@ -17,4 +45,4 @@
     <p style="text-align:center; font-weight:bold; font-size:20px">Please select an option from the left hand menu.
 </div>
 
-<?php include("../includes/footer.php"); ?>
+<?php include("../includes/footer.php"); }?>
