+++ otherPages/loggingin.php
@@ -10,8 +10,8 @@
   }
   
   if($_POST['psw'] == 'abc123' && $_POST['uname'] == 'admin') {
-    $_SESSION['loggedin'] = true;
-    $_SESSION['role'] = 'admin';
+    $_SESSION['logged_in'] = true;
+    $_SESSION['role'] = 'ADMIN';
     $_SESSION['user'] = $_POST['uname'];
     $_SESSION['email'] = 'placeholder@example.com';
     header("location:login.php?status=loggedin");
